Update: EOSLobbyManager, EOSTransport, EOSEACLobbyManager: Enhanced functionality of existing Callbacks, Added additional Voice Channel related functions and changed accessibility of Lookup functions related to Client Identities
Class: EOSLobbyManager.cs
Affected Class: EOSEACLobbyManager.cs
Add new LobbyChangedEvents params to OnLobbyChanged Callbacks so all Clients will now be capable of identifying what triggered a LobbyChanged callback.

Added new MuteLocalMember and DeafenLocalMember functions which do not rely on CurrentLobby.Members List. The existing ToggleMute function relies on the CurrentLobby.Members list to be populated and will not work before that update is received. MuteLocalMember and DeafenLocalMember can be execute as soon as the Lobby is created.

Added new MuteMember function that will mute a Target ProductUserId. It is similar to the existing ToggleMute function just that this one isn't a toggle and will set Mute status based of the specified value.

Class: EOSTransport.cs
Modified Accessibility (Changed to Public) of the following functions:
public ProductUserId GetUserId(ulong clientId)
public ulong GetClientId(ProductUserId userId)
The above 2 functions are useful for other custom classes to build upon, especially for Host Migration where we need to be able to lookup the corresponding ProductUserID and ClientID.